+++ NEWS.md
@@ -1,6 +1,4 @@
-# R4GoodPersonalFinances (development version)
-
-# R4GoodPersonalFinances 0.0.0.9000
+# R4GoodPersonalFinances 0.0.0.9016
 
 Initial release of the package includes functions designed to 
 support the optimal allocation of your investment portfolio
@@ -9,3 +7,5 @@ with the use of the 'Merton Share' formula.
 * Added `calc_purchasing_power()` and `plot_purchasing_power()` to calculate and visualize the effect of real interest rates—whether positive (compounding) or negative (inflation)—on the purchasing power of savings.  
 * Added `calc_optimal_risky_asset_allocation()` for calculating optimal allocation using the 'Merton Share' formula.  
 * Added `calc_risk_adjusted_return()` and `plot_risk_adjusted_returns()` to calculate and visualize the effect of allocation to risky assets on the risk-adjusted return.
+* Added `run_app()` function that runs package interactive apps:
+`risk-adjusted-returns` and `purchasing-power`.